Поправил формат данных списаний

+++ internal/model/balance.go
@@ -2,5 +2,5 @@ package model
 
 type Balance struct {
 	Current   float64 `json:"current"`
-	Withdrawn int     `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 }
